refactor(core): share the History-family epoch math
Extract the duplicated 1601-us `base::Time` conversion and SQLite
integer-range helpers into a single pure `forensic-time.ts`, consumed by
both `history.ts` and `downloads.ts`. Downloads live inside the History
database and share History's exact epoch, so the offset constant,
`floorDivision`, `utcFromUnixMicros`, and `boundedSortInteger` now have
one authoritative home instead of byte-for-byte copies. Analyzer output
is unchanged.
