+++ core/src/history.ts
@@ -33,6 +33,11 @@ import {
 } from "./case.js";
 import { ForensixError, WorkingCopyIntegrityRefusal } from "./errors.js";
 import {
+  boundedSortInteger,
+  utcFromUnixMicros,
+  WINDOWS_EPOCH_OFFSET_MICROS,
+} from "./forensic-time.js";
+import {
   absentField,
   createFinding,
   unavailableField,
@@ -249,9 +254,6 @@ interface BuiltVisit {
   readonly provenance: Provenance;
 }
 
-const WINDOWS_EPOCH_OFFSET_MICROS = 11_644_473_600_000_000n;
-const SQLITE_MAX_INTEGER = 9_223_372_036_854_775_807n;
-const SQLITE_MIN_INTEGER = -9_223_372_036_854_775_808n;
 const TRANSITION_CORES = new Map<number, string>([
   [0, "link"],
   [1, "typed"],
@@ -358,28 +360,6 @@ function preservedString(
   return typeof value === "string"
     ? valueField(value)
     : unavailableField("unsupported_value");
-}
-
-function floorDivision(value: bigint, divisor: bigint): bigint {
-  const quotient = value / divisor;
-  const remainder = value % divisor;
-  return remainder < 0n ? quotient - 1n : quotient;
-}
-
-function utcFromUnixMicros(unixMicros: bigint): string | null {
-  const seconds = floorDivision(unixMicros, 1_000_000n);
-  const micros = unixMicros - seconds * 1_000_000n;
-  const milliseconds = seconds * 1000n + micros / 1000n;
-  const numericMilliseconds = Number(milliseconds);
-  if (!Number.isSafeInteger(numericMilliseconds)) {
-    return null;
-  }
-  const date = new Date(numericMilliseconds);
-  if (Number.isNaN(date.getTime())) {
-    return null;
-  }
-  const base = date.toISOString();
-  return `${base.slice(0, -5)}.${micros.toString().padStart(6, "0")}Z`;
 }
 
 function resolveHistoryEpoch(
@@ -892,12 +872,6 @@ function durationSummaryFields(visits: readonly BuiltVisit[]): {
         ? absentField()
         : valueField((total / BigInt(withDurationCount)).toString()),
   };
-}
-
-function boundedSortInteger(value: bigint): bigint | null {
-  return value >= SQLITE_MIN_INTEGER && value <= SQLITE_MAX_INTEGER
-    ? value
-    : null;
 }
 
 function buildSummaries(
